move scroll logic into useHasScrolled hook
